+++ docs/how-to-guides/how-to-run-the-cmip7-scenariomip-workflow.py
@@ -624,7 +624,7 @@ scm_runner = CMIP7ScenarioMIPSCMRunner.from_cmip7_scenariomip_config(
 # You will likely want to skip this step if running yourself.
 
 # %% editable=true slideshow={"slide_type": ""}
-if os.environ.get("READTHEDOCS", False):
+if os.environ.get("READTHEDOCS", "False") == "True":
     scm_runner.climate_models_cfgs["MAGICC7"] = scm_runner.climate_models_cfgs[
         "MAGICC7"
     ][:10]